requirements.txt
pip~=23.2.1
attrs~=24.2.0
wheel~=0.41.2
setuptools~=68.2.0
discord~=2.3.2
discord.py~=2.4.0
pytz~=2024.2
pymongo~=4.10.1
python-dotenv~=1.0.1